refactor: rename types into cosmos_types
