+++ packages/bindings-test/src/subspaces/msg.rs
@@ -4,9 +4,11 @@ mod tests {
     use crate::consts::{TEST_SUBSPACE, TEST_SUBSPACE_USER_GROUP, USER2_ADDRESS};
     use chrono::DateTime;
     use cosmwasm_std::{Addr, Coin};
+    use desmos_bindings::cosmos_types::{
+        Allowance, AuthzGrant, BasicAllowance, GenericAuthorization,
+    };
     use desmos_bindings::subspaces::msg::SubspacesMsg;
     use desmos_bindings::subspaces::types::{Grantee, Permission, UserGrantee};
-    use desmos_bindings::types::{Allowance, AuthzGrant, BasicAllowance, GenericAuthorization};
     use test_contract::msg::ExecuteMsg;
     use test_contract::msg::ExecuteMsg::DesmosMessages;
