+++ packages/bindings/src/reports/mocks.rs
@@ -1,10 +1,10 @@
 //! Contains useful mocks of the Desmos x/reports module's types made to be used in any test.
 
+use crate::cosmos_types::Timestamp;
 use crate::reports::types::{
     QueryReasonResponse, QueryReasonsResponse, QueryReportResponse, QueryReportsResponse, Reason,
     Report, UserTarget,
 };
-use crate::types::Timestamp;
 
 use chrono::DateTime;
