+++ packages/bindings/src/subspaces/mocks.rs
@@ -1,5 +1,6 @@
 //! Contains useful mocks of the Desmos x/subspaces module's types made to be used in any test.
 
+use crate::cosmos_types::{Allowance, BasicAllowance, Timestamp};
 use crate::subspaces::types::Permission;
 use crate::subspaces::types::{
     permission_detail::{Sum, User},
@@ -8,7 +9,6 @@ use crate::subspaces::types::{
     QueryUserAllowancesResponse, QueryUserGroupMembersResponse, QueryUserGroupResponse,
     QueryUserGroupsResponse, QueryUserPermissionsResponse, Section, Subspace, UserGroup,
 };
-use crate::types::{Allowance, BasicAllowance, Timestamp};
 use chrono::DateTime;
 
 /// Represents the mock subspace owner for unit test.
